+++ src/loggers/index.ts
@@ -5,6 +5,7 @@ export {
   logEventBridgeEvent,
   logS3Event,
   logDynamoDBStreamEvent,
+  logAppSyncEvent,
 } from './logLambdaEvent';
 
 export type { LogConfig } from './logLambdaEvent';